+++ packages/hooks/src/useAddFund.ts
@@ -3,6 +3,7 @@ import { gql } from 'graphql-request'
 import { useCallback, useContext, useState } from 'react'
 import invariant from 'ts-invariant'
 import { LiteflowContext } from './context'
+import useConfig from './useConfig'
 import { ErrorMessages } from './errorMessages'
 
 gql`
@@ -19,7 +20,10 @@ export default function useAddFund(
 ): [() => Promise<void>, { loading: boolean }] {
   const { sdk } = useContext(LiteflowContext)
   const [loading, setLoading] = useState(false)
+  const config = useConfig()
+
   const addFunds = useCallback(async () => {
+    invariant((await config).hasTopUp, ErrorMessages.FEATURE_DISABLED_TOP_UP)
     invariant(signer, ErrorMessages.SIGNER_FALSY)
     try {
       setLoading(true)
@@ -32,6 +36,6 @@ export default function useAddFund(
     } finally {
       setLoading(false)
     }
-  }, [sdk, signer])
+  }, [sdk, signer, config])
   return [addFunds, { loading }]
 }