+++ packages/hooks/src/useInvitation.ts
@@ -3,6 +3,7 @@ import { gql } from 'graphql-request'
 import { useCallback, useContext, useState } from 'react'
 import invariant from 'ts-invariant'
 import { LiteflowContext } from './context'
+import useConfig from './useConfig'
 import { ErrorMessages } from './errorMessages'
 
 gql`
@@ -40,9 +41,14 @@ export default function useInvitation(signer: Signer | undefined): {
   creating: boolean
 } {
   const { sdk } = useContext(LiteflowContext)
+  const config = useConfig()
   const [accepting, setAccepting] = useState(false)
   const [creating, setCreating] = useState(false)
   const create = useCallback(async () => {
+    invariant(
+      (await config).hasReferralSystem,
+      ErrorMessages.FEATURE_DISABLED_REFERRAL,
+    )
     invariant(signer, ErrorMessages.SIGNER_FALSY)
     try {
       setCreating(true)
@@ -61,10 +67,14 @@ export default function useInvitation(signer: Signer | undefined): {
     } finally {
       setCreating(false)
     }
-  }, [sdk, signer])
+  }, [sdk, signer, config])
 
   const accept = useCallback(
     async (invitationId: string) => {
+      invariant(
+        (await config).hasReferralSystem,
+        ErrorMessages.FEATURE_DISABLED_REFERRAL,
+      )
       try {
         setAccepting(true)
         const { acceptInvitation } = await sdk.AcceptInvitation({
@@ -79,7 +89,7 @@ export default function useInvitation(signer: Signer | undefined): {
         setAccepting(false)
       }
     },
-    [sdk],
+    [sdk, config],
   )
 
   return { create, accept, accepting, creating }