+++ packages/hooks/CHANGELOG.md
@@ -6,6 +6,13 @@
 
 #### Added
 
+- Add `useConfig` hook to return the configuration of the platform [#65](https://github.com/liteflow-labs/libraries/pull/65)
+  - `hasLazyMint` is true when lazymint is activated
+  - `hasReferralSystem` is true when the referral and invitation system is activated
+  - `hasSocialFeatures` is true when all the social features (likes, comments...) are activated
+  - `hasTopUp` is true when fiat on ramp is activated
+  - `hasUnlockableContent` is true when unlockable content is activated
+
 #### Changed
 
 #### Deprecated
@@ -13,6 +20,8 @@
 #### Removed
 
 #### Fixed
+
+- Add proper errors when calling a hook relying on a feature not activated [#65](https://github.com/liteflow-labs/libraries/pull/65)
 
 #### Security