+++ packages/templates/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### Breaking Changes
 
+- `AssetForm` template does not require the `activateLazyMint` and `activateUnlockableContent` props anymore [#65](https://github.com/liteflow-labs/libraries/pull/65)
+
 #### Added
 
 #### Changed